move instantiation to module map
